Set up buttons and new game logic in console
This commit will set up the game to be played via buttons rather than
prompt, and will change a game to be played until a score of 5 is
reached for either side.

Clean up needed, or a restructured game() function should be written as
it is not used in this commit.